Fix errors and add mypy to CI (#2190)
* Remove adjoint, unused and untested.

* Remove this, cannot work with missing RTLD_NOW in modern ctypes.

* Work in progress for mypy pass (MetaClass stuff is tricky).

* Add mypy to ubuntu image

* Again

* Move later

* Bit more typing, remove adjoint import

* Fix but more typing errors back

* Adding if makes type check pass

* 35 errors left in python/dolfinx

* Remaining issues are due to use of single dispatch

* Add py.typed files so that users actually see type annotations

* Generalised arguments types for mesh constructors

* Fix singledispatch definition to raise NotImplementedError if no method found.

* Fixes all typing issues in assemble.

* 9 left in demos, 22 in main library (mainly in petsc.py)

* flake8 fixes

* Demos pass, petsc.py and two strange ones to go

* Liberal mypy passes, all tests pass

* Fix fallback vector

* Fixes

* mypy settings in cfg file

* Small tweak

* Bound numpy while 1.22 breaks numba

* pep8

* Remove numpy version

* Move back to old use

* fix

* flake8

* Test fixes

* move mypy up

* fix mypy install

* Fixes

* isort

* Free numpy version

* Might need to re-run tests once Docker images update

* Add mypy to image

* Type pybind11 .

It seems like there are projects to generate typing data from pybind11 automagically
but I think it looks too complex for a few methods.

* Remove mypy install from CI

* Also wrap cell_perm_gmsh

* Make dim optional, grab tdim?

* Reallow value to be passed

* Allow lists to eval

* Remove comment

* Better API?

* Prefer this design.

* Better singledispatch la

* Bug fixes

* Fix

* Ignore return error

* flake8

* Remove explicit casts

* Remove cryptic typings, ignore errors

* Revert some files to main

* Remove explicit dim

* Small tweaks.

* Revert

* flake8

* Fix, should discuss this again

+++ python/demo/demo_gmsh.py
@@ -32,6 +32,7 @@ from dolfinx.io import (XDMFFile, cell_perm_gmsh, distribute_entity_data,
 from dolfinx.mesh import CellType, create_mesh, meshtags_from_entities
 
 from mpi4py import MPI
+
 # -
 
 # Generate a mesh on each rank with the gmsh API, and create a DOLFINx
@@ -119,7 +120,7 @@ msh.name = "ball_d1"
 entities, values = distribute_entity_data(msh, 2, marked_facets, facet_values)
 
 msh.topology.create_connectivity(2, 0)
-mt = meshtags_from_entities(msh, 2, create_adjacencylist(entities), np.int32(values))
+mt = meshtags_from_entities(msh, 2, create_adjacencylist(entities), values)
 mt.name = "ball_d1_surface"
 
 with XDMFFile(MPI.COMM_WORLD, "out_gmsh/mesh.xdmf", "w") as file:
@@ -179,7 +180,7 @@ marked_facets = marked_facets[:, gmsh_triangle6]
 
 entities, values = distribute_entity_data(msh, 2, marked_facets, facet_values)
 msh.topology.create_connectivity(2, 0)
-mt = meshtags_from_entities(msh, 2, create_adjacencylist(entities), np.int32(values))
+mt = meshtags_from_entities(msh, 2, create_adjacencylist(entities), values)
 mt.name = "ball_d2_surface"
 with XDMFFile(MPI.COMM_WORLD, "out_gmsh/mesh.xdmf", "a") as file:
     file.write_mesh(msh)
@@ -256,7 +257,7 @@ msh.name = "hex_d2"
 
 entities, values = distribute_entity_data(msh, 2, marked_facets, facet_values)
 msh.topology.create_connectivity(2, 0)
-mt = meshtags_from_entities(msh, 2, create_adjacencylist(entities), np.int32(values))
+mt = meshtags_from_entities(msh, 2, create_adjacencylist(entities), values)
 mt.name = "hex_d2_surface"
 
 with XDMFFile(MPI.COMM_WORLD, "out_gmsh/mesh.xdmf", "a") as file:
